Allow the user to supply multiple ports per broadcast address.
This changes the format of the backend configuration.
To allow for an easy transition of old configs we react on the failure
to convert the config to a dict and try an automatic conversion to the
new format. If this succeeds the user will be given a proper warning
about the changed format and will also receive his converted config that
he then can put into his config.